OpenLGTV BCM: fm.cgi and fm-action.cgi improvements - buttons to copy/move in background, average kbps check, better elapsed time check, button to cancel copy/move

+++ OpenLGTV_BCM-src/var/www/fm-action.cgi
@@ -6,6 +6,7 @@ Content-type: text/html
 <HEAD>
 
 <!-- OpenLGTV BCM FileManager by xeros -->
+<!-- fm-action.cgi script for handling actions like copy/move/delete/play and confirm/cancel of operation -->
 <!-- Source code released under GPL License -->
 
 <style type="text/css">
@@ -215,9 +216,13 @@ if [ "$side" = "l" ]
 then
     export spth="$lpth"
     export dpth="$rpth"
+    export lpthx="`dirname $lpth`"
+    export rpthx="$rpth"
 else
     export spth="$rpth"
     export dpth="$lpth"
+    export lpthx="$lpth"
+    export rpthx="`dirname $rpth`"
 fi
 
 #echo "spth: $spth dpth: $dpth lpth: $lpth rpth: $rpth"
@@ -226,6 +231,28 @@ echo '<center><font size="+3" color="yellow"><br/><br/><b>OpenLGTV BCM FileManag
 
 #echo '<div style="position: absolute; left: 10px; top: 10px; width:860px; font-size:16px; background-color:white;">'
 echo '<div style="width:80%; margin:auto; font-size:16px; background-color:white;">'
+
+[ "$FORM_pid" != "" ] && pid="$FORM_pid"
+
+if [ -n "$FORM_cancel" -a -n "$pid" ]
+then
+    process="`ps | grep \"^ *$pid \"`"
+    if [ -n "$process" ]
+    then
+	echo "<font size='+3' color='black'><br/><b>Cancelling process:</b><br/><br/>$process<br/><br/>"
+	kill "$pid" 2>&1
+	sleep 3
+	process="`ps | grep \"^ *$pid \"`"
+	if [ -n "$process" ]
+	then
+	    kill -9 "$pid" 2>&1
+	fi
+    fi
+    #echo "side: $side lpthx: $lpthx rpthx: $rpthx"
+    #sleep 5
+    echo "<script type='text/javascript'>window.location='fm.cgi?type=related&side=$side&lpth=$lpthx&rpth=$rpthx';</script>"
+    echo "</font></div></center></body></head></html>"
+fi
 
 if [ "$action" = "play" ]
 then
@@ -263,34 +290,42 @@ then
     echo "</b><br/><br/></font>"
     [ "$action" != "play" ] && echo "<table><tr><td id='tr_l1' width='200px' align='center'><b><a id='link_l1' href='${REQUEST_URI}&confirm=yes'><font size='+4'>Yes</font></a></b></td><td id='tr_r1' width='200px' align='center'><b><a id='link_r1' href='javascript:history.go(-1);'><font size='+4'>No</font></a></b></td></tr></table></center><br/><br/>"
 else
-    if [ "$action" = "copy" ]
+    if [ "$action" = "copy" -o "$action" = "move" ]
     then
-	echo "Copying $spth to $dpth/ ...<br/><br/>"
-	if [ "$FORM_onlystatus" != "1" ]
-	then
-	    cp -r "$spth" "$dpth/" > /var/log/cp.log 2>&1 &
-	    pid="$!"
-	    date +"%s" > /var/log/cp.date.${pid}.log
-	else
-	    [ "$FORM_pid" != "" ] && pid="$FORM_pid"
-	fi
-	time_start="`cat /var/log/cp.date.${pid}.log`"
-	echo "<div id='status' style='font-size: 30px;'></div>"
-	dfile="`basename "$spth"`"
-	# TODO: make dialog with buttons for operations - 'copy in background' and 'cancel'
-	echo "<table><tr><td id='tr_l1' width='200px' align='center'><b><a id='link_l1' href='${REQUEST_URI}&confirm=yes'><font size='+4'>Yes</font></a></b></td><td id='tr_r1' width='200px' align='center'><b><a id='link_r1' href='javascript:history.go(-2);'><font size='+4'>No</font></a></b></td></tr></table></center><br/><br/>"
-	counter=0
-	#sleep 1
+	echo "$action $spth to $dpth/ ..." | tee -a /var/log/${action}.log
+	echo "<br/><br/>"
 	SIFS="$IFS" IFS=$'\n'
 	ssize=$(for i in `find "$spth" ! -type d`; do stat -c "%s" "$i"; done | awk '{sum += $1} END{print sum}') # could have been done with '-printf "%s\n"' or '-exec stat -c "%s" {}' as find arguments but busybox find does not support properly both of them
 	IFS="$SIFS"
-	[ "$ssize" = "" ] && ssize=1
+	if [ "$FORM_onlystatus" != "1" ]
+	then
+	    if [ "$action" = "copy" ]
+	    then
+		cp -r "$spth" "$dpth/" > /var/log/${action}.log 2>&1 &
+		pid="$!"
+	    else
+		mv "$spth" "$dpth/" > /var/log/${action}.log 2>&1 &
+		pid="$!"
+	    fi
+	    date +"%s" > /var/log/${action}.date.${pid}.log
+	fi
+	time_start="`cat /var/log/${action}.date.${pid}.log`"
+	time_start_status="`date +'%s'`"
+	echo "<div id='status' style='font-size: 30px;'></div>"
+	dfile="`basename "$spth"`"
+	#echo "<table><tr><td id='tr_l1' width='500px' align='center'><b><a id='link_l1' href='javascript:history.go(-2);'><font size='+4'>Continue in background</font></a></b></td><td id='tr_r1' width='300px' align='center'><b><a id='link_r1' href='${REQUEST_URI}&pid=${pid}&cancel=1'><font size='+4'>Cancel</font></a></b></td></tr></table></center><br/><br/>"
+	echo "<table><tr><td id='tr_l1' width='500px' align='center'><b><a id='link_l1' href='fm.cgi?type=related&side=$side&lpth=$lpthx&rpth=$rpthx'><font size='+4'>Continue in background</font></a></b></td><td id='tr_r1' width='300px' align='center'><b><a id='link_r1' href='${REQUEST_URI}&pid=${pid}&cancel=1'><font size='+4'>Cancel</font></a></b></td></tr></table></center><br/><br/><br/>"
+	counter=0
+	[ -z "$ssize" ] && ssize=1
+	sleep 1
+	#echo "ssize: $ssize"
 	for i in `seq 2000`
 	do
 	    #ssize=$(busybox stat -c "%s " "$spth")        # only files
 	    #dsize=$(busybox stat -c "%s" "$dpth/$dfile")
 	    #ssize=$(du -s "$spth" | cut -f1)              # in KB, dirs and files but on one filesystem - different values on different fs
 	    #dsize=$(du -s "$dpth/$dfile" | cut -f1)
+	    #echo "i: $i"
 	    SIFS="$IFS" IFS=$'\n'
 	    dsize=$(for i in `find "$dpth/$dfile" ! -type d`; do stat -c "%s" "$i"; done | awk '{sum += $1} END{print sum}') 
 	    IFS="$SIFS"
@@ -299,8 +334,11 @@ else
 	    percent="$(($dsize * 100 / $ssize))"
 	    time_now="`date +'%s'`"
 	    elapsed=$((${time_now}-${time_start}))
+	    elapsed_status=$((${time_now}-${time_start_status}))
+	    average_bps=$((${dsize}/${elapsed}))
+	    average_kbps=$((${average_bps}/1024))
 	    #echo "<script type='text/javascript'>document.getElementById('status').innerHTML ='Copied: $dsize / $ssize bytes<br/><br/>Progress: $percent%<br/><br/>Elapsed time: $counter seconds<br/><br/>';</script>"
-	    echo "<script type='text/javascript'>document.getElementById('status').innerHTML ='Copied: $dsize / $ssize bytes<br/><br/>Progress: $percent%<br/><br/>Elapsed time: $elapsed seconds<br/><br/>';</script>"
+	    echo "<script type='text/javascript'>document.getElementById('status').innerHTML ='Copied: $dsize / $ssize bytes<br/><br/>Progress: $percent% &nbsp; Average speed: $average_kbps kbps<br/><br/>Elapsed time: $elapsed seconds<br/><br/>';</script>"
 	    #echo "Copied: $dsize / $ssize bytes<br/><br/>Progress: $percent%<br/><br/>"
 	    if [ -z "`ps | grep \"^ *$pid \"`" ]
 	    then
@@ -309,14 +347,18 @@ else
 		    echo "Finished"
 		    break
 		else
-		    echo "<font color='red' size='+3'><b>ERROR copying file!</b><br/><br/>"
-		    [ -f "/var/log/cp.log" ] && cat /var/log/cp.log
-		    echo "</font>"
+		    echo "<font color='red' size='+3'><b>ERROR copying file!</b></font><br/><br/>"
+		    if [ -f "/var/log/${action}.log" ]
+		    then
+			echo "<font color='red' size='+2'>"
+			cat /var/log/${action}.log
+			echo "</font>"
+		    fi
 		    error=1
 		    break
 		fi
 	    fi
-	    if [ "$counter" -gt "120" ]
+	    if [ "${elapsed_status}" -gt "120" ]
 	    then
 		if [ "$pid" != "" ]
 		then
@@ -326,18 +368,8 @@ else
 		fi
 	    fi
 	    sleep 3
-	    counter=$(($counter+3))
+	    counter=$(($counter+1))
 	done
-    fi
-    if [ "$action" = "move" ]
-    then
-	echo "Moving $spth to $dpth/ ...<br/><br/>"
-	mv "$spth" "$dpth/" 2>&1
-	#cp -r "$spth" "$dpth/" 2>&1
-	if [ "$?" -ne "0" ]
-	then
-	    error=1
-	fi
     fi
     if [ "$action" = "delete" ]
     then
@@ -353,10 +385,10 @@ else
     fi
     if [ "$error" != "1" ]
     then
-	echo "<br/><br/><font color='green' size='+2'><b>DONE</b></font><br/>"
+	echo "<br/><br/><font color='green' size='+3'<b>DONE</b></font><br/>"
 	timeout=2000
     else
-	echo "<br/><br/><font color='red' size='+3'><b>ERROR</b></font><br/>"
+	echo "<br/><br/><font color='red' size='+4'><b>ERROR</b></font><br/>"
 	timeout=10000
     fi
     #sleep 2
